+++ src/seamlesserverpkg/renderer/base.nim
@@ -5,9 +5,10 @@ import seamlesserverpkg/renderer/base/[
   footer,
   flash,
   rendered,
-  state
+  state,
+  bridgedData
 ]
-export header, footer, flash, rendered, state
+export header, footer, flash, rendered, state, bridgedData
 
 import seamlesserverpkg/renderer/base/karaxNodes/dynamicLink
 export dynamicLink
@@ -19,7 +20,7 @@ when not defined js:
   import pkg/prologue except appName
 
   import seamlesserverpkg/config
-  import seamlesserverpkg/renderer/base/bridgedData
+  from seamlesserverpkg/auth/utils import isLogged
 
   type Render = proc(state: State): Rendered
 
@@ -27,8 +28,10 @@ when not defined js:
     ## Server side rendering of Karax model
     var state = ctx.newState
     for flash in ctx.getFlashedMsgsWithCategory():
-      state.brData.flash(parseEnum[FlashLevel](flash[0]), flash[1])
-    state.brData.flash(Info, "Hellow from server! this message will disappear after 2s", 2000)
+      state.brData.flash(flash[1], parseEnum[FlashLevel](flash[0]))
+    state.brData.flash("Hellow from server! this message will disappear after 2s", Info, 2000)
+    state.brData.isLogged = ctx.isLogged
+    echo state.brData.isLogged
     withConf:
       let rendered = render state
       let
@@ -39,11 +42,13 @@ when not defined js:
         title: text rendered.genTitle appName
         meta(name = "viewport", content = "width=device-width, -scale=1.0")
       body:
+        rendered.ssrvnodes.before
         tdiv(id = "ROOT"):
           renderHeader state
           renderFlashes state.brData.flashes
           rendered.vnode
           renderFooter state
+        rendered.ssrvnodes.after
         genBridgedData state.brData
         script(src = "/" & jsDir / jsFile)
     result = $vnode