+++ src/seamlesserverpkg/renderer/base/flash.nim
@@ -1,7 +1,6 @@
 from std/times import getTime, toUnix
 
 include pkg/karax/prelude
-from pkg/prologue/core/types import FlashLevel
 
 import seamlesserverpkg/renderer/base/bridgedData
 
@@ -18,9 +17,9 @@ proc renderFlashes*(flashes): VNode =
 
 proc flash*(
   brData: var BridgedData;
-  level: FlashLevel;
   text: string;
-  duration = 1000 # ms
+  level: FlashLevel;
+  duration = 5000 # ms
 ) =
   ## Creates new flash message in BridgedData
   brData.flashes.add FlashMessage(
@@ -30,13 +29,28 @@ proc flash*(
     duration: duration
   )
 
+proc del*(mflashes; text: string; all = false): bool =
+  ## Deletes flashes with provided text message and returns true if succeeds
+  result = false
+  var i = 0
+  for _ in 0..<mflashes.len:
+    if mflashes[i].text == text:
+      mflashes.del i
+      result = true
+      dec i
+      if not all:
+        break
+    inc i
 when defined js:
   from std/dom import setTimeout
 
   proc autoDeleteFlashes*(mflashes) =
     ## Automatically delete the flashes after they expire
-    for i, flash in mflashes:
-      discard setTimeout(proc =
-        mflashes.del i
-        redraw()
-      , flash.duration)
+    echo mflashes
+    for flash in mflashes.mitems:
+      if not flash.deleting:
+        flash.deleting = true
+        discard setTimeout(proc =
+          discard mflashes.del flash.text
+          redraw()
+        , flash.duration)