+++ src/seamlesserverpkg/renderer/routes/user/login.nim
@@ -1,12 +1,21 @@
 include pkg/karax/prelude
 
 import seamlesserverpkg/renderer/base
+from seamlesserverpkg/userMessages import umAlreadyLoggedIn
+import seamlesserverpkg/renderer/base/flash
+
+when defined js:
+  from seamlesserverpkg/js/utils import redirect
 
 const path* = "/user/login"
 
 proc renderHtml*(state: State): Rendered =
   ## Login page HTML multi-backend renderer
-  new result
+  result = newRendered()
+  when defined js:
+    if state.brData.isLogged:
+      state.brData.flash(umAlreadyLoggedIn, Warning)
+      redirect "/"
   result.title = "Login"
   result.vnode = buildHtml(main):
     h1: text "Login page"
@@ -22,18 +31,22 @@ proc renderHtml*(state: State): Rendered =
         `type` = "password"
       )
       button(`type` = "submit"): text "Submit"
-
-    
+  when not defined js:
+    result.ssrvnodes.before = buildHtml(tdiv):
+      span: text "Hi, this is a static server-side rendered HTML before JS main node"
+    result.ssrvnodes.after = buildHtml(tdiv):
+      span: text "Hi, this is a static server-side rendered HTML after JS main node"
 
 when not defined js:
   import pkg/prologue
 
-  import seamlesserverpkg/renderer/utils
-  
+  import seamlesserverpkg/auth/utils
+
   proc get*(ctx: Context) {.async.} =
     ## GET login page
     doAssert ctx.request.reqMethod == HttpGet
-    let user = ctx.loggedUser
-    if user.isNil:
-      discard redirect "/"
+    if ctx.isLogged:
+      ctx.flash(umAlreadyLoggedIn, Warning)
+      resp redirect "/"
+      return
     resp ctx.ssr renderHtml