+++ src/seamlesserverpkg/userMessages.nim
@@ -1,4 +1,11 @@
 const
-  umInvalidEmail* = "Invalid email"
+  umInvalidEmail* = "Invalid email."
   umInvalidPassword* = "Invalid password. Make sure that's 6 characters up to 32."
   umInvalidUsername* = "Invalid username. Make sure that's smaller than 16 characters."
+  umRegisterSuccess* = "User successfully created!"
+  umUserExists* = "Username was already taken."
+  umEmailExists* = "Email was already used."
+  umRegisterError* = "An error occurred, if it persists, please contact an admin."
+    ## Unknown register error 
+  umAlreadyLoggedIn* = "You already logged in"
+  umErrorParseBridgedData* = "Error to parse the server's data: '$#'"